// projects/codx-ac/src/lib/models/Vendors.model.ts
export class Vendors {
  accountStatement: any;
  address: any;
  allowEditKey: any;
  alterVendor: any;
  assign: any;
  autoSettlement: any;
  balAmt: any;
  buid: any;
  buyerGroupID: any;
  calendarControl: any;
  cashDiscID: any;
  channelID: any;
  checkCreditLimit: any;
  city: any;
  contactID: any;
  contactName: any;
  countryID: any;
  createdBy: any;
  createdOn: any;
  creditBalance: any;
  creditLimit: any;
  creditLimitControl: any;
  creditRate: any;
  cumulateGroupID: any;
  currencyID: any;
  customerID: any;
  debtComparision: any;
  delModeID: any;
  delRouteID: any;
  delTermID: any;
  delete: any;
  distance: any;
  districtID: any;
  email: any;
  establishYear: any;
  exchangeRate: any;
  faxNo: any;
  freightChargeGroupID: any;
  id: any;
  idField: any;
  includeTables: any;
  industries: any;
  industryID: any;
  invoiceAccount: any;
  invoiceForm: any;
  invoiceSeriNo: any;
  jobTitle: any;
  latitude: any;
  leadtime: any;
  lineDiscGroupID: any;
  longitude: any;
  mixedPayment: any;
  mobile: any;
  modifiedBy: any;
  modifiedOn: any;
  multiDiscGroupID: any;
  noOfEmployees: any;
  note: any;
  orderDeadline: any;
  orderDuplicateDays: any;
  orderLeadtime: any;
  overDueControl: any;
  overDueDays: any;
  overDueInvoices: any;
  overdueAmt: any;
  owner: any;
  phone: any;
  phoneExt: any;
  pmtMethodID: any;
  pmtSchedID: any;
  pmtTermID: any;
  postDetail: any;
  postDetailItems: any;
  postItems: any;
  postalCode: any;
  priceGroupID: any;
  promotionGroupID: any;
  provinceID: any;
  purchasePoolID: any;
  quotationDeadline: any;
  recID: any;
  regionID: any;
  salesConsultantID: any;
  salespersonID: any;
  searchName: any;
  segmentID: any;
  settleInvoice: any;
  settlePayment: any;
  settlement: any;
  share: any;
  sms: any;
  statisticGroupID: any;
  status: any;
  stop: any;
  stopOn: any;
  stopReason: any;
  street: any;
  timeZone: any;
  totalDiscGroupID: any;
  unbounds: any;
  updateColumns: '';
  vendorFullName: any;
  vendorGroupID: any;
  vendorID: any;
  vendorName: any;
  vendorName2: any;
  vendorTaxCode: any;
  warehouseID: any;
  webPage: any;
  workingDays: any;
  workingHour: any;
  write: any;
}